requirements.txt
anyio==4.3.0
certifi==2024.2.2
dnspython==2.6.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
psutil==5.9.8
pymongo==4.6.2
python-dotenv==1.0.1
python-telegram-bot==21.0.1
sniffio==1.3.1
typing_extensions==4.10.0
